travel_concierge/sub_agents/inspiration/pyproject.toml
[project]
name = "inspiration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
    "a2a-sdk>=0.3.11",
    "google-adk>=1.18.0",
]
